Fix: fixed handle blackjack events to not exit early right after initialise hands, changed deck to be bytes instead and encoded onto the blockchain

+++ contracts/Interfaces/IOracle.sol
@@ -3,5 +3,5 @@ pragma solidity ^0.8.30; // may have to change for ganache
 import "contracts/Data.sol";
 
 interface IOracle {
-    function generateSeed(uint256 playerSeed) external view returns (uint);
+    function generateSeed(uint256 playerSeed) external view returns ( uint256 );
 }
